+++ js/fs.js
@@ -128,7 +128,7 @@ var nfs = module.exports = {
      */
     list: function(path)
     {
-        if (path !== "" && !path.startsWith("/")) return Promise.reject(new Error("Wield workspace path "+path));
+        if (path !== "" && !path.startsWith("/")) return Promise.reject(new Error("Path must starts with slash: "+path));
         return callbackToPromise((callback)=>fs.readdir(nfs.workspace + path, callback));
     },
     /**
@@ -137,7 +137,7 @@ var nfs = module.exports = {
      */
     stat: function(path)
     {
-        if (path !== "" && !path.startsWith("/")) return Promise.reject(new Error("Wield workspace path "+path));
+        if (path !== "" && !path.startsWith("/")) return Promise.reject(new Error("Path must starts with slash: "+path));
         return callbackToPromise((callback)=>fs.stat(nfs.workspace + path, callback));
     },
     /**
@@ -146,7 +146,7 @@ var nfs = module.exports = {
      */
     mkdir: function(path)
     {
-        if (!path.startsWith("/")) return Promise.reject(new Error("Wield workspace path "+path));
+        if (!path.startsWith("/")) return Promise.reject(new Error("Path must starts with slash: "+path));
         return new Promise(function(resolve, reject){
             fs.mkdir(nfs.workspace + path, function(err){
                 if (err)
@@ -164,7 +164,7 @@ var nfs = module.exports = {
      */
     mkdirp: function(path)
     {
-        if (!path.startsWith("/")) return Promise.reject(new Error("Wield workspace path "+path));
+        if (!path.startsWith("/")) return Promise.reject(new Error("Path must starts with slash: "+path));
         return callbackToPromise((callback)=>mkdirParent(nfs.workspace + path, callback));
     },
     /**
@@ -173,7 +173,7 @@ var nfs = module.exports = {
      */
     lstat: function(path)
     {
-        if (path !== "" && !path.startsWith("/")) return Promise.reject(new Error("Wield workspace path "+path));
+        if (path !== "" && !path.startsWith("/")) return Promise.reject(new Error("Path must starts with slash: "+path));
         return callbackToPromise((callback)=>fs.lstat(nfs.workspace + path, callback));
     },
     /**
@@ -182,7 +182,7 @@ var nfs = module.exports = {
      */
     open: function(path)
     {
-        if (!path.startsWith("/")) return Promise.reject(new Error("Wield workspace path "+path));
+        if (!path.startsWith("/")) return Promise.reject(new Error("Path must starts with slash: "+path));
         return callbackToPromise((callback)=>fs.readFile(nfs.workspace + path, "utf-8", callback));
     },
 
@@ -192,7 +192,7 @@ var nfs = module.exports = {
      */
     exists: function(path)
     {
-        if (!path.startsWith("/")) return Promise.reject(new Error("Wield workspace path "+path));
+        if (!path.startsWith("/")) return Promise.reject(new Error("Path must starts with slash: "+path));
         return new Promise((resolve) => fs.exists(nfs.workspace + path, resolve));
     },
 
@@ -223,7 +223,7 @@ var nfs = module.exports = {
      */
     createSync: function(path, data)
     {
-        if (!path.startsWith("/")) throw new Error("Wield workspace path "+path);
+        if (!path.startsWith("/")) throw new Error("Path must starts with slash: "+path);
         return fs.writeFileSync(nfs.workspace + path, data, "utf-8");
     },
 
@@ -233,7 +233,7 @@ var nfs = module.exports = {
      */
     delete: function(path)
     {
-        if (!path.startsWith("/")) return Promise.reject(new Error("Wield workspace path "+path));
+        if (!path.startsWith("/")) return Promise.reject(new Error("Path must starts with slash: "+path));
         return callbackToPromise((callback)=>fs.unlink(nfs.workspace + path, callback));
     },
 
@@ -251,7 +251,7 @@ var nfs = module.exports = {
                 data[p] = defaultValue[p];
                 changed = true;
             }
-            if (!changed) return;
+            if (!changed) return data;
             return nfs.create(filepath, JSON.stringify(data, null, 4))
             .then(()=> data);
         })